+++ teneto/__init__.py
@@ -4,13 +4,13 @@
 """Teneto is a module with tools for analyzing network patterns in time."""
 #
 __author__ = "William Hedley Thompson (wiheto)"
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 #
 from . import networkmeasures
 from . import utils
 from . import plot
 from . import generatenetwork
-from . import derive
+from . import timeseries
 from . import misc
 from . import trajectory
 #from . import communitydetection